+++ pipeline/merger/src/main/scala/weco/pipeline/merger/services/Merger.scala
@@ -246,8 +246,8 @@ object PlatformMerger extends Merger {
           target,
           sources
         ).redirectSources
-        sourceImageData <- ImageDataRule(target, sources).redirectSources
-        redirectImageData <- ImagesRule(target, sources).redirectSources
+        targetImageData <- ImageDataRule(target, sources).redirectSources
+        separateImageData <- ImagesRule(target, sources).redirectSources
         work = target
           .mapData {
             data =>
@@ -255,12 +255,12 @@ object PlatformMerger extends Merger {
                 items = items,
                 thumbnail = thumbnail,
                 otherIdentifiers = otherIdentifiers,
-                imageData = sourceImageData
+                imageData = targetImageData
               )
           }
       } yield MergeResult(
         mergedTarget = modifyInternalWorks(work, items),
-        imageDataWithSources = redirectImageData.map {
+        imageDataWithSources = separateImageData.map {
           imageData =>
             ImageDataWithSource(
               imageData = imageData,
